requirements.txt
# DalioBot Trading Powerhouse - Dependencies
# Core
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Market Data (Free)
yfinance>=0.2.30
fredapi>=0.5.0
pandas-datareader>=0.10.0

# Brokerage
alpaca-trade-api>=3.0.0
alpaca-py>=0.21.0

# Technical Analysis
ta>=0.11.0
ta-lib>=0.4.28;platform_system!="Windows"

# ML & Optimization
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
torch>=2.0.0

# Backtesting
vectorbt>=0.26.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0

# Options
mibian>=0.1.3
py_vollib>=1.0.1

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
pyyaml>=6.0.0
loguru>=0.7.0
